+++ libft/srcs/ft_numlen.c
@@ -6,15 +6,15 @@
 /*   By: abarnett <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/26 10:14:30 by abarnett          #+#    #+#             */
-/*   Updated: 2018/06/06 16:49:23 by abarnett         ###   ########.fr       */
+/*   Updated: 2018/11/06 14:34:01 by abarnett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int			ft_numlen(long num)
+size_t		ft_numlen(long num)
 {
-	int	length;
+	size_t	length;
 
 	length = 1;
 	if (num < 0)
